Crate netherrack

Entry module for the Netherrack project

Modules

core
Core functionality for Netherrack

io
Access to various I/O operations and features

universe
The universe, where everything to do with server-wide functions resides

Structs

Receiver
The receiving-half of Rust's channel type. This half can only be owned by one thread

Sender
The sending-half of Rust's asynchronous channel type. This half can only be owned by one thread, but it can be cloned to send to other threads.

Version
Represents a version number conforming to the semantic versioning scheme.

Constants

MINECRAFT_PROTOCOL_VERSION
The version of the Minecraft protocol accepted by Netherrack

NETHERRACK_VERSION_STRING
The version of Netherrack as determined at compile time

Functions

channel
Creates a new asynchronous channel, returning the sender/receiver halves.

get_version
Gets the current version of Netherrack

start_server
Starts a new Netherrack server instance

Implementations