Struct netherrack::io::network::game_connection::GameConnection

pub struct GameConnection {
    pub stream: TcpStream,
    pub state: ConnectionState,
    pub forge_enabled: bool,
    pub connected: bool,
}

A struct defining a game connection

Fields

stream

The TCP stream bound to this GameConnection

state

The state of this GameConnection

forge_enabled

Whether the client has the Forge Mod Loader injected into it

This may change in 1.9 as FML is being destroyed and Forge is becoming the norm

connected

A boolean declaring if this GameConnection is connected

Methods

impl GameConnection

fn new(stream: TcpStream) -> GameConnection

Creates a new GameConnection

fn disconnect(&mut self)

fn start_listening(&mut self)

Starts listening on the connection